requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
grpcio>=1.41.0
grpcio-tools>=1.41.0
protobuf>=3.19.0
pydantic>=1.8.2
typing-extensions>=4.0.0
